Add refresh() command before confirm stone removal
During multiplayer play, some text elements were not being displayed to
the screen due to a missing refresh() call. I also noticed that some
messages are getting truncated at the bottom when the screen is too
short to display all of the text. This might warrent an overhaul of
displaying text.