Fix bug with stone capturing
This was a tough bug to find, and then fix. Basically stones that were
set to be removed were added to a stack, and also a list. The list was
to make sure any future stones couldn't be added to the stack for
removal. However I only added the points to the list after popping them
from the list, which meant that if a point was pushed to the stack, then
another one of those points was pushed to the stack before popping the
original point, it would be added twice to the list which was used for
counting how many stones were removed at the end. This resulted in an
incorrect stone count in very rare circumstances.