Add naive implementation of resign command.
The limitations with what I have added are that anyone can use the
resign command at any time during networked play and it will always
count the resign as coming from the player who's turn it is.
Additionally when playing against an engine, there is no way to stop the
engine from interacting with the game after the game is over and it
enters edit/view mode.