Add comment support
I modularized the code for displaying certain things to the screen, so
now the board, info, and comments all are handled seperately. This also
means that now certain aspects of the display can be disabled. The "say"
command was also changed so that it adds a comment instead of just
printing to the notes buffer.