View beginning of game after load SGF
The boardview is now set to the beginning of the game after loading an
SGF file, which is more in line with other SGF viewing programs. It also
fixes a bug where the tree view was messed up due to not tracking
view_index during gametree creation while loading the sgf.